+++ lib/oli/delivery/sections/section.ex
@@ -69,6 +69,8 @@ defmodule Oli.Delivery.Sections.Section do
     has_many :section_project_publications, SectionsProjectsPublications, on_replace: :delete
 
     field :enrollments_count, :integer, virtual: true
+    field :total_count, :integer, virtual: true
+    field :institution_name, :string, virtual: true
 
     timestamps(type: :utc_datetime)
   end