+++ lib/oli_web/live/common/paged_table.ex
@@ -8,8 +8,8 @@ defmodule OliWeb.Common.PagedTable do
   prop limit, :integer, required: true
   prop offset, :integer, required: true
   prop table_model, :struct, required: true
-  prop sort, :event, required: true
-  prop page_change, :event, required: true
+  prop sort, :event, default: "paged_table_sort"
+  prop page_change, :event, default: "paged_table_page_change"
 
   def render(assigns) do
     ~F"""
@@ -26,5 +26,44 @@ defmodule OliWeb.Common.PagedTable do
       {/if}
       </div>
     """
+  end
+
+  def handle_delegated(event, params, socket, patch_fn, model_key \\ :table_model) do
+    delegate_handle_event(event, params, socket, patch_fn, model_key)
+  end
+
+  def delegate_handle_event("paged_table_page_change", %{"offset" => offset}, socket, patch_fn, _) do
+    patch_fn.(socket, %{offset: offset})
+  end
+
+  # handle change of selection
+  def delegate_handle_event(
+        "paged_table_sort",
+        %{"sort_by" => sort_by},
+        socket,
+        patch_fn,
+        model_key
+      ) do
+    sort_order =
+      case Atom.to_string(socket.assigns[model_key].sort_by_spec.name) do
+        ^sort_by ->
+          if socket.assigns[model_key].sort_order == :asc do
+            :desc
+          else
+            :asc
+          end
+
+        _ ->
+          socket.assigns[model_key].sort_order
+      end
+
+    patch_fn.(socket, %{
+      sort_by: sort_by,
+      sort_order: sort_order
+    })
+  end
+
+  def delegate_handle_event(_, _, _, _) do
+    :not_handled
   end
 end